+++ bugsnag-android-core/src/test/java/com/bugsnag/android/EventErrorTypeTest.kt
@@ -1,6 +1,5 @@
 package com.bugsnag.android
 
-import junit.framework.TestCase.assertNull
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import java.io.File
@@ -32,7 +31,7 @@ class EventErrorTypeTest {
     fun verifyLegacyFileNameErrorType() {
         val config = convertToImmutableConfig(BugsnagTestUtils.generateConfiguration())
         val file = File("1504255147933_0000111122223333aaaabbbbcccc9999_my-uuid-123.json")
-        val payload = EventPayload(config.apiKey, file, Notifier(), config)
+        val payload = EventPayload(config.apiKey, null, file, Notifier(), config)
         assertEquals(emptySet<ErrorType>(), payload.getErrorTypes())
     }
 
@@ -43,7 +42,7 @@ class EventErrorTypeTest {
     fun verifyFileNameDefault() {
         val config = convertToImmutableConfig(BugsnagTestUtils.generateConfiguration())
         val file = File("1504255147933_0000111122223333aaaabbbbcccc9999_android_my-uuid-123_.json")
-        val payload = EventPayload(config.apiKey, file, Notifier(), config)
+        val payload = EventPayload(config.apiKey, null, file, Notifier(), config)
         assertEquals(setOf(ErrorType.ANDROID), payload.getErrorTypes())
     }
 
@@ -54,7 +53,7 @@ class EventErrorTypeTest {
     fun verifyFileNameMixedTypes() {
         val config = convertToImmutableConfig(BugsnagTestUtils.generateConfiguration())
         val file = File("1504255147933_0000111122223333aaaabbbbcccc9999_android,c,reactnativejs_my-uuid-123_.json")
-        val payload = EventPayload(config.apiKey, file, Notifier(), config)
+        val payload = EventPayload(config.apiKey, null, file, Notifier(), config)
         assertEquals(ErrorType.values().toSet(), payload.getErrorTypes())
     }
 }