+++ features/fixtures/mazerunner/cxx-scenarios-bugsnag/src/main/cpp/cxx-scenarios-bugsnag.cpp
@@ -277,4 +277,12 @@ Java_com_bugsnag_android_mazerunner_scenarios_CXXBackgroundNotifyScenario_activa
   bugsnag_notify_env(env, (char *)"Ferret Escape!", (char *)"oh no", BSG_SEVERITY_ERR);
 }
 
+JNIEXPORT void JNICALL
+Java_com_bugsnag_android_mazerunner_scenarios_CXXNotifySmokeScenario_activate(JNIEnv *env,
+                                                                              jobject instance) {
+  bugsnag_set_user_env(env, (char *)"324523", NULL, (char *)"Jack Mill");
+  bugsnag_leave_breadcrumb_env(env, (char *)"Cold beans detected", BSG_CRUMB_LOG);
+  bugsnag_notify_env(env, (char *)"CXXNotifySmokeScenario", (char *)"Smoke test scenario", BSG_SEVERITY_ERR);
+}
+
 }